the final commit

+++ task3.c
@@ -2,16 +2,16 @@
 
 int main(){
 
-    int ang1, ang2, ang3, angT;
+    float ang1, ang2, ang3, angT;
 
     printf("Enter angle 1: ");
-    scanf("%d", &ang1);
+    scanf("%f", &ang1);
 
     printf("Enter angle 2: ");
-    scanf("%d", &ang2);
+    scanf("%f", &ang2);
 
     angT= ang1 + ang2;
 
-    printf("Third angle is %d", ang3 = 180 - angT);
+    printf("Third angle is %f", ang3 = 180 - angT);
 
 }
